incremental version ...

+++ user/pg/ddl/create.sql
@@ -13,8 +13,8 @@ CREATE TABLE if not exists users (
     settings_avatar_description_value text not null default '',
     settings_avatar_description_image text not null default '',
     settings_radar_perimeter int not null default 10000,
-    settings_radar_position_latitude int not null default 0,
-    settings_radar_position_longitude int not null default 0
+    settings_radar_position_latitude double precision not null default 0,
+    settings_radar_position_longitude double precision not null default 0
 );
 
 create table if not exists user_filesystem (
